refactor: 깔금한 코드를 위해 Racing class 만들어서 controller 역할 분리
